Update Plugins
Akismet Anti-Spam
Alter White Label Admin
Amazon Web Services
Appointments+
Avatars For Multisite
bbPress
Beaver Builder Plugin (Pro Version)
Black Studio TinyMCE Widget
BNE Testimonials
BuddyPress
Calls to Action
Cloudflare
CoursePress Pro
Custom Sidebars Pro
E-Newsletter
Flatty - UI Admin Theme
Google Analytics +
HTML Editor Syntax Highlighter
HubSpot Marketing Free (formerly Leadin)
Hustle Pro
Image Widget
Jetpack by WordPress.com
Landing Pages
Leads
MailChimp Sync
MailPoet Newsletters
MarketPress
Master Slider Pro
mb.YTPlayer for background videos
Members
Ninja Forms
Optimize Database after Deleting Revisions
Pay Per View
Reveal IDs
Shareaholic
Shortcodes Ultimate
Snapshot
Social Metrics Tracker
Support System
Ultimate Branding
Ultimate Member
WooCommerce
WordPress Chat
WPShapere
WP Admin UI
WP Defender Pro
WP Hummingbird
WP Migrate DB Pro
WPMU DEV Dashboard
WP Offload S3
WP Offload S3 - Assets Addon
WP Smush Pro

+++ content/plugins/wp-hummingbird/core/modules/class-module-minify.php
@@ -7,6 +7,8 @@ require_once( 'minify/class-minify-sources-collector.php' );
 // New files
 include_once( 'minify/class-minify-group.php' );
 include_once( 'minify/class-minify-groups-list.php' );
+include_once( 'minify/class-minify-housekeeper.php' );
+include_once( 'minify/class-minify-scanner.php' );
 
 class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 
@@ -28,6 +30,16 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 	public $errors_controller;
 
 	/**
+	 * @var WP_Hummingbird_Module_Minify_Housekeeper
+	 */
+	public $housekeeper;
+
+	/**
+	 * @var WP_Hummingbird_Module_Minify_Scanner
+	 */
+	public $scanner;
+
+	/**
 	 * Counter that will name scripts/styles slugs
 	 *
 	 * @var int
@@ -47,6 +59,10 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 
 	public function __construct( $slug, $name ) {
 		parent::__construct( $slug, $name );
+		$this->housekeeper = new WP_Hummingbird_Module_Minify_Housekeeper();
+		$this->housekeeper->init();
+
+		$this->scanner = new WP_Hummingbird_Module_Minify_Scanner();
 		self::$counter = 0;
 	}
 
@@ -66,6 +82,9 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 		add_filter( 'wp_hummingbird_is_active_module_' . $this->get_slug(), array( $this, 'should_be_active' ), 20 );
 
 		add_action( 'before_delete_post', array( $this, 'on_delete_post' ), 10 );
+
+		// Process the queue through WP Cron
+		add_action( 'wphb_minify_process_queue', array( $this, 'process_queue' ) );
 	}
 
 	/**
@@ -74,8 +93,8 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 	public function on_delete_post( $post_id ) {
 		$group = WP_Hummingbird_Module_Minify_Group::get_instance_by_post_id( $post_id );
 
-		if ( is_a( $group, 'WP_Hummingbird_Module_Minify_Group'  ) && $group->file_id ) {
-			if ( $group->get_file_path() ) {
+		if ( ( $group instanceof WP_Hummingbird_Module_Minify_Group ) && $group->file_id ) {
+			if ( $group->get_file_path() && file_exists( $group->get_file_path() ) ) {
 				wp_delete_file( $group->get_file_path() );
 			}
 			wp_cache_delete( 'wphb_minify_groups' );
@@ -83,41 +102,11 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 	}
 
 	public function should_be_active( $is_active ) {
-		if ( ! $this->can_execute_php() ) {
+		if ( ! wphb_can_execute_php() ) {
 			return false;
 		}
 
 		return $is_active;
-	}
-
-	/**
-	 * Check if Hummingbird is currently checking files
-	 *
-	 * @return bool
-	 */
-	public static function is_checking_files() {
-		$checking_files = get_option( 'wphb-minification-check-files', 0 );
-		if ( $checking_files )
-			return true;
-
-		return false;
-	}
-
-	/**
-	 * Check if the current PHP version is suitable for minification
-	 */
-	public function can_execute_php() {
-		$minimun = $this->get_php_min_version();
-
-		if ( version_compare( PHP_VERSION, $minimun, '<' ) ) {
-			return false;
-		}
-
-		return true;
-	}
-
-	public function get_php_min_version() {
-		return '5.3';
 	}
 
 	public function run() {
@@ -125,7 +114,7 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 
 		add_action( 'init', array( $this, 'register_cpts' ) );
 
-		if ( is_admin() || is_customize_preview() || is_a( $wp_customize, 'WP_Customize_Manager' ) ) {
+		if ( is_admin() || is_customize_preview() || ( $wp_customize instanceof WP_Customize_Manager ) ) {
 			return;
 		}
 
@@ -136,7 +125,7 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 		//add_action( 'wp_head', array( $this, 'print_scripts' ), 900 );
 		//add_action( 'wp_print_footer_scripts', array( $this, 'print_late_resources' ), 900 );
 
-		add_action( 'shutdown', array( $this, 'process_queue' ) );
+		add_action( 'wp_footer', array( $this, 'trigger_process_queue_cron' ), 10000 );
 	}
 
 	/**
@@ -240,10 +229,12 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 
 			// If we aren't in footer, remove handles that need to go to footer
 			if ( ! self::is_in_footer() && isset ( $wp_dependencies->registered[ $handle ]->extra['group'] ) && $wp_dependencies->registered[ $handle ]->extra['group'] ) {
+				$this->to_footer[$type][] = $handle;
 				unset( $handles[ $key ] );
-				$this->to_footer[ $type ][] = $handle;
 			}
 		}
+
+		$handles = array_values( $handles );
 
 		$return_to_wp = array();
 
@@ -255,19 +246,11 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 		// Group dependencies by attributes like args, extra, etc
 		$_groups = $this->group_dependencies_by_attributes( $handles, $wp_dependencies, $type );
 
-
 		// Create a Groups list object
 		$groups_list = new WP_Hummingbird_Module_Minify_Groups_List( $type );
 		array_map(array( $groups_list, 'add_group' ), $_groups );
 
 		unset( $_groups );
-
-		$this->attach_inline_attribute( $groups_list, $wp_dependencies );
-
-		if ( 'scripts' === $type ) {
-			$this->attach_scripts_localization( $groups_list, $wp_dependencies );
-		}
-
 
 		// Time to split the groups if we're not combining some of them
 		foreach ( $groups_list->get_groups() as $group ) {
@@ -285,57 +268,14 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 			}
 
 			$dont_combine_list = $group->get_dont_combine_list();
-
 			if ( $dont_combine_list ) {
-				// There are one or more handles that should not be combined
-				/** @var WP_Hummingbird_Module_Minify_Group $group */
-				$handles = $group->get_handles();
-				// Here we'll save sources that don't need to be minified/combine
-				// Then we'll extract those handles from the group and we'll create
-				// a new group for them keeping the groups order
-				$group_combine = array();
-				foreach ( $handles as $handle ) {
-					$combine_resource = $group->should_do_handle( $handle, 'combine' );
-					$group_combine[ $handle ] = $combine_resource ? 1 : 0;
-				}
+				$splitted_group = $this->_get_splitted_group_structure_by( 'combine', $group );
+				// Split the group!
+				$groups_list->split_group( $group->hash, $splitted_group );
+			}
 
-				// Now split groups if needed based on combine value
-				// We need to keep always the order, ALWAYS
-				// This will save the new splitted group structure
-				$splitted_group = array();
-
-				$last_status = null;
-				foreach ( $group_combine as $handle => $status ) {
-
-					// Last minify status will be the first one by default
-					if ( is_null( $last_status ) ) {
-						$last_status = $status;
-					}
-
-					// Set the splitted groups to the last element
-					end( $splitted_group );
-					if ( $last_status === $status && $status !== 0 ) {
-						$current_key = key( $splitted_group );
-						if ( ! $current_key ) {
-							// Current key can be NULL, set to 0
-							$current_key = 0;
-						}
-
-						if ( ! isset( $splitted_group[ $current_key ] ) || ! is_array( $splitted_group[ $current_key ] ) ) {
-							$splitted_group[ $current_key ] = array();
-						}
-
-						$splitted_group[ $current_key ][] = $handle;
-					}
-					else {
-						// Create a new group
-						$splitted_group[] = array( $handle );
-					}
-
-
-					$last_status = $status;
-				}
-
+			if ( 'scripts' === $type && $group->get_defer_list() ) {
+				$splitted_group = $this->_get_splitted_group_structure_by( 'defer', $group, false );
 				// Split the group!
 				$groups_list->split_group( $group->hash, $splitted_group );
 			}
@@ -356,10 +296,14 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 			}
 		}
 
+		if ( 'scripts' === $type ) {
+			$this->attach_scripts_localization( $groups_list, $wp_dependencies );
+		}
+		$this->attach_inline_attribute( $groups_list, $wp_dependencies );
+
 		// Parse dependencies, load files and mark groups as ready,process or only-handles
 		// Watch out! Groups must not be changed after this point
 		$groups_list->preprocess_groups();
-
 
 		foreach ( $groups_list->get_groups() as $group ) {
 			$group_status = $groups_list->get_group_status( $group->hash );
@@ -383,12 +327,75 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 					if ( $group->should_process_group() ) {
 						$this->group_queue[] = $group;
 					}
-
 				}
 			}
 		}
 
 		return $return_to_wp;
+	}
+
+	/**
+	 * Create a new group structure based on $by parameter
+	 *
+	 * This will allow later to split groups into new groups based on combination/deferring...
+	 *
+	 * @param string $by combine|defer|minify...
+	 * @param WP_Hummingbird_Module_Minify_Group $group
+	 * @param bool $value Value to apply if the handle should be done
+	 *
+	 * @return array New structure
+	 */
+	private function _get_splitted_group_structure_by( $by, $group, $value = true ) {
+		$handles = $group->get_handles();
+
+		// Here we'll save sources that don't need to be minified/combine/deferred...
+		// Then we'll extract those handles from the group and we'll create
+		// a new group for them keeping the groups order
+		$group_todos = array();
+		foreach ( $handles as $handle ) {
+			$value = absint( $value );
+			$not_value = absint( ! $value );
+			$group_todos[ $handle ] = $group->should_do_handle( $handle, $by ) ? $value : $not_value;
+		}
+
+		// Now split groups if needed based on $by value
+		// We need to keep always the order, ALWAYS
+		// This will save the new splitted group structure
+		$splitted_group = array();
+
+		$last_status = null;
+		foreach ( $group_todos as $handle => $status ) {
+
+			// Last minify status will be the first one by default
+			if ( is_null( $last_status ) ) {
+				$last_status = $status;
+			}
+
+			// Set the splitted groups to the last element
+			end( $splitted_group );
+			if ( $last_status === $status && $status !== 0 ) {
+				$current_key = key( $splitted_group );
+				if ( ! $current_key ) {
+					// Current key can be NULL, set to 0
+					$current_key = 0;
+				}
+
+				if ( ! isset( $splitted_group[ $current_key ] ) || ! is_array( $splitted_group[ $current_key ] ) ) {
+					$splitted_group[ $current_key ] = array();
+				}
+
+				$splitted_group[ $current_key ][] = $handle;
+			}
+			else {
+				// Create a new group
+				$splitted_group[] = array( $handle );
+			}
+
+
+			$last_status = $status;
+		}
+
+		return $splitted_group;
 	}
 
 	/**
@@ -482,6 +489,8 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 
 				// We'll treat this later
 				$new_group->delete_extra( 'after' );
+				$new_group->delete_extra( 'before' );
+				$new_group->delete_extra( 'data' );
 
 				$new_group->set_args( $registered_dependency->args );
 
@@ -532,13 +541,18 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 		$registered = $wp_dependencies->registered;
 		$extras = wp_list_pluck( $registered, 'extra' );
 		$after = wp_list_pluck( array_filter( $extras, array( $this, '_filter_after_after_attribute' ) ), 'after' );
+		$before = wp_list_pluck( array_filter( $extras, array( $this, '_filter_after_before_attribute' ) ), 'before' );
 
-		array_map( function( $group ) use ( $groups_list, $after ) {
+		array_map( function( $group ) use ( $groups_list, $after, $before ) {
 			/** @var WP_Hummingbird_Module_Minify_Group $group */
-			array_map( function( $handle ) use ( $after, $group ) {
+			array_map( function( $handle ) use ( $after, $group, $before ) {
 				if ( isset( $after[ $handle ] ) ) {
 					// Add!
 					$group->add_after( $after[ $handle ] );
+				}
+				if ( isset( $before[ $handle ] ) ) {
+					// Add!
+					$group->add_before( $before[ $handle ] );
 				}
 			}, $group->get_handles() );
 		}, $groups_list->get_groups() );
@@ -583,6 +597,18 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 		return false;
 	}
 
+	/**
+	 * Filter a list of dependencies returning their 'before' attribute inside 'extra' list
+	 *
+	 * @internal
+	 */
+	public function _filter_after_before_attribute( $a ) {
+		if ( isset ( $a['before'] ) ) {
+			return $a['before'];
+		}
+		return false;
+	}
+
 
 	/**
 	 * Return if we are processing the footer
@@ -594,32 +620,52 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 	}
 
 	/**
-	 * Process the queue: Minify and combine files
+	 * Trigger the action to process the queue
 	 */
-	public function process_queue() {
-		$queue = $this->get_queue_to_process();
+	public function trigger_process_queue_cron() {
+		// Trigger que the queue hrough WP CRON so we don't waste load time
 		$this->sources_collector->save_collection();
 
+		$queue = $this->get_queue_to_process();
+		$this->add_items_to_persistent_queue( $queue );
+		$queue = $this->get_pending_persistent_queue();
 		if ( empty( $queue ) ) {
 			return;
 		}
 
-		$this->add_items_to_persistent_queue( $queue );
+		if ( defined('DISABLE_WP_CRON' ) && DISABLE_WP_CRON ) {
+			$this->process_queue();
+		}
+		else {
+			self::schedule_process_queue_cron();
+		}
+	}
 
+	/**
+	 * Process the queue: Minify and combine files
+	 */
+	public function process_queue() {
 		// Process the queue
 		if ( get_transient( 'wphb-processing' ) ) {
-			// Still processing
+			// Still processing. Try again
+			if ( ! ( defined('DISABLE_WP_CRON' ) && DISABLE_WP_CRON ) ) {
+				self::schedule_process_queue_cron();
+			}
 			return;
 		}
 
+		$queue = $this->get_pending_persistent_queue();
+
 		set_transient( 'wphb-processing', true, 60 );
-		// Process 4 groups max in a request
+		// Process 10 groups max in a request
 		$count = 0;
+
+		$new_queue = $queue;
 		foreach ( $queue as $key => $item ) {
-			if ( $count >= 4 ) {
+			if ( $count >= 8 ) {
 				break;
 			}
-			if ( ! is_a( $item, 'WP_Hummingbird_Module_Minify_Group' ) )
+			if ( ! ( $item instanceof WP_Hummingbird_Module_Minify_Group ) )
 				continue;
 
 			/** @var WP_Hummingbird_Module_Minify_Group $item */
@@ -630,9 +676,25 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 				}
 			}
 			$this->remove_item_from_persistent_queue( $item->hash );
+			unset( $new_queue[ $key ] );
 			$count++;
 		}
+
+		if ( ! ( defined('DISABLE_WP_CRON' ) && DISABLE_WP_CRON ) ) {
+			$new_queue = array_values( $new_queue );
+			if ( ! empty( $new_queue ) ) {
+				// Still needs processing
+				self::schedule_process_queue_cron();
+			}
+		}
+
 		delete_transient( 'wphb-processing' );
+	}
+
+	public static function schedule_process_queue_cron() {
+		if ( ! wp_next_scheduled( 'wphb_minify_process_queue' ) ) {
+			wp_schedule_single_event( time(), 'wphb_minify_process_queue' );
+		}
 	}
 
 	/**
@@ -644,11 +706,25 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 	 * @param array $items
 	 */
 	private function add_items_to_persistent_queue( $items ) {
+		if ( empty( $items ) ) {
+			// Nothing to be added
+			return;
+		}
 		$current_queue = $this->get_pending_persistent_queue();
 		if ( empty( $current_queue ) ) {
-			add_option( 'wphb_process_queue', $items, null, false );
+			update_option( 'wphb_process_queue', $items );
 		} else {
-			update_option( 'wphb_process_queue', array_merge( $items, $current_queue ) );
+			$updated = false;
+			$current_queue_hashes = wp_list_pluck( $current_queue, 'hash' );
+			foreach ( $items as $item ) {
+				if ( ! in_array( $item->hash, $current_queue_hashes ) ) {
+					$updated = true;
+					$current_queue[] = $item;
+				}
+			}
+			if ( $updated ) {
+				update_option( 'wphb_process_queue', $current_queue );
+			}
 		}
 	}
 
@@ -696,7 +772,7 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 	/**
 	 * Clear the module cache
 	 */
-	public static function clear_cache( $reset_settings = true ) {
+	public function clear_cache( $reset_settings = true ) {
 		global $wpdb;
 
 		// Clear all the cached groups data
@@ -718,6 +794,7 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 			delete_option( $name );
 		}
 
+
 		WP_Hummingbird_Sources_Collector::clear_collection();
 
 		wphb_minification_clear_files();
@@ -736,7 +813,7 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 		// Clear the pending process queue
 		self::clear_pending_process_queue();
 
-		update_option( 'wphb-minification-check-files', 0 );
+		$this->scanner->reset_scan();
 
 		WP_Hummingbird_Minification_Errors_Controller::clear_errors();
 	}
@@ -744,107 +821,6 @@ class WP_Hummingbird_Module_Minify extends WP_Hummingbird_Module {
 	public static function clear_pending_process_queue() {
 		delete_option( 'wphb_process_queue' );
 		delete_transient( 'wphb-processing' );
-	}
-
-	/**
-	 * Initializes the scanning process
-	 */
-	public static function init_scan() {
-		wphb_clear_minification_cache( false );
-
-		// Activate minification if is not
-		wphb_toggle_minification( true );
-
-		// Calculate URLs to Check
-		$args = array(
-			'orderby'        => 'rand',
-			'posts_per_page' => '1',
-			'ignore_sticky_posts' => true,
-			'post_status' => 'publish'
-		);
-
-		$urls = array();
-
-		$urls[] = home_url();
-
-		$post_types = get_post_types();
-		$post_types = array_diff( $post_types, array( 'attachment', 'nav_menu_item', 'revision' ) );
-
-		foreach ( $post_types as $post_type ) {
-			$args['post_type'] = $post_type;
-			$posts = get_posts( $args );
-			if ( $posts ) {
-				$urls[] = get_permalink( $posts[0] );
-			}
-
-			$post_type_archive_link = get_post_type_archive_link( $post_type );
-			if ( $post_type_archive_link )
-				$urls[] = $post_type_archive_link;
-		}
-
-		if ( get_option( 'show_on_front' ) && $post = get_post( get_option( 'page_for_posts' ) ) ) {
-			$urls[] = get_permalink( $post->ID );
-		}
-
-		$urls = array_unique( $urls );
-
-		$urls_list = array();
-		// Duplicate every URL 4 times. This will be enough to generate all the files for most of the sites
-		for ( $i = 0; $i < 4; $i++ ) {
-			$urls_list = array_merge( $urls_list, $urls );
-		}
-
-		sort( $urls_list );
-		$urls = $urls_list;
-
-		$args = array(
-			'on' => current_time( 'timestamp' ),
-			'urls_list' => $urls,
-			'urls_done' => array()
-		);
-
-		update_option( 'wphb-minification-check-files', $args );
-	}
-
-	/**
-	 * This function send a request to a URL in the site
-	 * that will trigger the files collection
-	 *
-	 * Executed with AJAX
-	 *
-	 * @param string $url
-	 *
-	 * @return array
-	 */
-	public static function scan( $url ) {
-		$cookies = array();
-		foreach ( $_COOKIE as $name => $value ) {
-			if ( strpos( $name, 'wordpress_' ) > -1 ) {
-				$cookies[] = new WP_Http_Cookie( array( 'name' => $name, 'value' => $value ) );
-			}
-
-		}
-
-		$result = array();
-
-		$args = array(
-			'timeout' => 0.01,
-			'cookies' => $cookies,
-			'blocking' => false,
-			'sslverify' => false
-		);
-		$result['cookie'] = wp_remote_get( $url, $args );
-
-		// One call logged out
-		$args = array(
-			'timeout' => 0.01,
-			'blocking' => false,
-			'sslverify' => false
-		);
-
-		$result['no-cookie'] = wp_remote_get( $url, $args );
-
-		return $result;
 	}
 
 }
